Add audio sensor functionality to Main method
This commit enhances the `Main` method in the `Program` class within the `SENSORS` namespace. It introduces an `Audio` object, initializes a list of sensor names with "audio", and creates a list of `Sensor` objects containing the `Audio` instance. The `investRank1` method from the `Investigation` class is called with the sensor names and the list of sensors, followed by retrieving a sensor using the `getSensor` method of the `IranianAgent` class.